Move tests to persist/fs package.
