+++ persist/fs/equals.go
@@ -1,4 +1,4 @@
-package roundtriptest
+package fs
 
 import (
 	indexfield "github.com/xichen2020/eventdb/index/field"
